[#7] Testing Update…
Made changes to pass tests 1, 2 and 3 for Home page.  This is working quite well now and I have update the css to format the page to look like the live site, but with some changes to make more consistent (e.g. normal text is serif whereas on the live site some pages are serif and some aren't).

Still more testing with other pages and more diverse formatting before complete.