[#7] Good Progress.  Started to create the text page files for each of the pages.  In the process encountered a problem which was that in order for links to (and also the titles of) pages to look right it must be possible for them to have spaces.  But I don't want to put spaces in the text file names as this creates other issues.  So I made a small change to the page creation script which takes underscores in the page name and replaces them with spaces in the title of the page.  Later on I will do something similar when creating the top level link bar.
The text file name now serves quite a lot of functions and I quite like how much I have squeezed in to such a simple artifact!

The plan is to complete the creation of the text files and then test each page for format and browser compatibility.  At that point this ticket can be closed with confidence.