[#19:status=>closed] Although the script was behaving as required I wanted to ensure that the images are displayed in order of filename so that this can be used to change the order easily if required.
I used the Perl sort command (which took a bit of understanding!) to achieve this, using the following line of code:

    @image_array = sort { $a->[0] cmp $b->[0]} @image_array;

This seems to work fine.  I may need to do something to ensure that case differences between file names (or actually probably just in the extension - jpg vs JPG) don't make any difference.  But for now I can close this ticket.